Main functions

Spatial geometry utilities

These are the main utilities to build spatial discretization objects.

UtilitiesForMRI.spatial_geometry — Method

spatial_geometry(field_of_view::NTuple{3,T}, nsamples::NTuple{3,Integer};
                 origin::NTuple{3,T}=field_of_view./2) where {T<:Real}

Return a 3D spatial geometry object that summarizes a Cartesian spatial discretization. Must specify domain size and number of samples per dimension. The origin of the domain can be changed by setting the keyword origin.

Example:

X = spatial_geometry((1f0, 1f0, 1f0), (64, 64, 64); origin=(0f0, 0f0, 0f0))

$k$-space geometry utilities

These are the main utilities to specify $k$-space acquisition trajectories, where the ordering is typically dictated by the order of acquisition (e.g., a proxy for time).

UtilitiesForMRI.kspace_sampling — Method

kspace_sampling(permutation_dims::NTuple{3,Integer},
                coord_phase_encoding::AbstractArray{T,2},
                coord_readout::AbstractVector{T}) where {T<:Real}

Returns a 3D $k$-space trajectory that is (time-wise) ordered by phase-encoding plane coordinates. The plane is orthogonal to a specific coordinate direction (e.g. $x$, $y$, or $z$). The input permutation_dims specifies the phase-encoding plane dimensions (in this order) and readout dimension. The coord_phase_encoding should be an array of size $n_t\times 3$ and represents the phase-encoding plane coordinate. Similarly, the readout coordinate coord_readout should be an array of size $n_k$.

UtilitiesForMRI.kspace_sampling — Method

kspace_sampling(X::CartesianSpatialGeometry{T},
                phase_encoding_dims::NTuple{2,Integer};
                phase_encode_sampling::Union{Nothing,AbstractVector{<:Integer}}=nothing,
                readout_sampling::Union{Nothing,AbstractVector{<:Integer}}=nothing) where {T<:Real}

Returns a 3D Cartesian $k$-space trajectory that is (time-wise) ordered by phase-encoding plane coordinates. The plane is orthogonal to a specific coordinate direction (e.g. $x$, $y$, or $z$). The spatial discretization is determined by the CartesianSpatialGeometry object X. The phase_encoding_dims indicate the phase-encoding plane dimensions. Subsampling of the full Cartesian k-space is obtained with the optional keyword inputs phase_encode_sampling and readout_sampling.

Fourier-transform utilities

In order to manipulate the non-uniform Fourier operator based on rigid-motion perturbation, see Section Getting started .

UtilitiesForMRI.nfft_linop — Method

nfft_linop(X::CartesianSpatialGeometry, K::StructuredKSpaceSampling)

Return the non-uniform Fourier transform as a linear operator for a specified Cartesian spatial discretization X and a $k$-space trajectory K.

Example

X = spatial_geometry((1f0, 1f0, 1f0), (32, 32, 32))
K = kspace_sampling(X, (1, 2))
F = nfft_linop(X, K)
u = randn(ComplexF32, X.nsamples)
d = F*u  # evaluation
    F'*d # adjoint

Resampling utilities

We list the main functionalities for subsampling/upsampling of spatial geometries, $k$-space geometries, and 3D images:

UtilitiesForMRI.resample — Method

resample(X::CartesianSpatialGeometry, n::NTuple{3,Integer})

Up/down-sampling of Cartesian discretization geometry. The field of view and origin of X are kept the same, while the sampling is changed according to n.

UtilitiesForMRI.subsample — Method

subsample(K, k_max::Union{T,NTuple{3,T}}; radial::Bool=false)

Down-sampling of a $k$-space trajectory. The output is a subset of the original $k$-space trajectory such that the $k$-space coordinates $(k_1,k_2,k_3)$ are: - $k_i\le k_{\mathrm{max},i}$ (if radial=false), or - $||\mathbf{k}||\le k_{\mathrm{max}}$ (if radial=true). The ordering of the subsampled trajectory is inherited from the original trajectory.

UtilitiesForMRI.subsample — Method

subsample(K, X::CartesianSpatialGeometry;
          radial::Bool=false, also_readout::Bool=true)

Down-sampling of a $k$-space trajectory, similarly to subsample. The maximum cutoff frequency is inferred from the Nyquist frequency of a Cartesian spatial geometry X.

UtilitiesForMRI.subsample — Method

subsample(K, d::AbstractArray{<:Complex,2}, Kq; norm_constant=nothing)

Down-sampling of a $k$-space data array d associated to a $k$-space trajectory K, e.g. d_i=d(\mathbf{k}_i). The output is a subset of the original data array, which is associated to the down-sampled $k$-space Kq (obtained, for example, via subsample). The keyword norm_constant allows the rescaling of the down-sampled data.

UtilitiesForMRI.subsample — Method

subsample(K, d::AbstractArray{<:Complex,2}, Kq; norm_constant=nothing)

Down-sampling of a $k$-space data array d associated to a $k$-space trajectory K, e.g. d_i=d(\mathbf{k}_i). The output is a subset of the original data array, which is associated to the down-sampled $k$-space Kq (obtained, for example, via subsample). The keyword norm_constant allows the rescaling of the down-sampled data.

UtilitiesForMRI.resample — Method

resample(u, n_scale::NTuple{3,Integer}; damping_factor=nothing)

Resampling of spatial array u. The underlying field of view represented by u is maintained, while the original sampling rate n=size(u) is changed to n_scale.

Image-quality metrics

Convenience functions to compute slice-based PSNR and SSIM metrics for 3D images (these relies on the package ImageQualityMetrics ):

UtilitiesForMRI.psnr — Method

psnr(u_noisy::AbstractArray{T,3}, u_ref::AbstractArray{T,3};
     slices::Union{Nothing,NTuple{N,VolumeSlice}}=nothing,
     orientation::Orientation=standard_orientation()) where {T<:Real,N}

Compute 2D power signal-to-noise ratio for the indicated 2D slices of a 3D array. The optional keyword slices indicates the 2D slices in object (see volume_slice), according to the 3D orientation (see orientation).

UtilitiesForMRI.ssim — Method

ssim(u_noisy::AbstractArray{T,3}, u_ref::AbstractArray{T,3};
     slices::Union{Nothing,NTuple{N,VolumeSlice}}=nothing,
     orientation::Orientation=standard_orientation()) where {T<:Real,N}

Compute 2D structural similarity index for the indicated 2D slices of a 3D array. The optional keyword slices indicates the 2D slices in object (see volume_slice), according to the 3D orientation (see orientation).

Visualization tools

UtilitiesForMRI.orientation — Method

orientation(perm::NTuple{3,Integer}; reverse::NTuple{3,Bool}=(false,false,false))

Defines an orientation for a 3D image. It is mostly useful for standardizing plotting utilities. For example it can be used to reorder a 3D image such that in radiological terms, after reordering: - 1st dimension = left-right - 2nd dimension = posterior-anterior - 3rd dimension = inferior-superior

The input perm determines an ordering for the 3D dimensions, while the keyword reverse specifies which dimensions are negatively oriented.

UtilitiesForMRI.standard_orientation — Method

standard_orientation()

Returns the standard orientation (no reordering).

UtilitiesForMRI.volume_slice — Method

volume_slice(dim::Integer, n::Integer; window=nothing)

Specifies a 2D slice with respect to a 3D image: dim is the dimension orthogonal to the 2D slice, n is the position of the slice, and window indicates a portion of the 2D slice.

Important: volume slices are always defined with respect to a given orientation of the 3D object.

UtilitiesForMRI.select — Method

select(u::AbstractArray{T,3}, slice::VolumeSlice; orientation::Orientation=standard_orientation()) where {T<:Real}

Returns the specified 2D slice of a 3D image (according to a specified orientation)

UtilitiesForMRI.plot_volume_slices — Method

plot_volume_slices(u::AbstractArray{T,3};
                   slices::Union{Nothing,NTuple{N,VolumeSlice}}=nothing,
                   spatial_geometry::Union{Nothing,CartesianSpatialGeometry{T}}=nothing,
                   cmap::String="gray",
                   vmin::Union{Nothing,Real}=nothing, vmax::Union{Nothing,Real}=nothing,
                   xlabel::Union{Nothing,AbstractString}=nothing, ylabel::Union{Nothing,AbstractString}=nothing,
                   cbar_label::Union{Nothing,AbstractString}=nothing,
                   title::Union{Nothing,AbstractString}=nothing,
                   savefile::Union{Nothing,String}=nothing,
                   orientation::Orientation=standard_orientation()) where {T<:Real,N}

Plot 2D slices of a given 3D image.